Demo final(sin deploy)

+++ api/src/models/Admin.js
@@ -32,11 +32,11 @@ module.exports = (sequelize) => {
     CALIFICACION: {
       type: DataTypes.ARRAY(DataTypes.FLOAT),
       allowNull: false,
-      defaultValue: [],
+      defaultValue: [], 
     },
 
     CELULAR: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.BIGINT,
     },
 
     FAVORITOS: {
